pyproject.toml
[project]
name = "pictex"
version = "0.3.0"
description = "A library to generate stylized text images."
authors = [
    { name = "Franco Zanardi", email = "francozanardi97@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["skia", "text-to-image", "image-generation", "typography", "graphics", "text-rendering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Natural Language :: English",
    "Topic :: Text Processing :: Fonts",
    "Typing :: Typed"
]

dependencies = [
  "skia-python",
  "typing_extensions >=4.0; python_version < '3.11'"
]

[project.urls]
Homepage = "https://github.com/francozanardi/pictex"
Repository = "https://github.com/francozanardi/pictex"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-regressions",
    "Pillow"
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]"
]
